--- src/main/groovy/org/openmicroscopy/blitz/BasePlugin.groovy
@@ -1,38 +1,0 @@
-/*
- * -----------------------------------------------------------------------------
- *  Copyright (C) 2019 University of Dundee & Open Microscopy Environment.
- *  All rights reserved.
- *
- *  This program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2 of the License, or
- *  (at your option) any later version.
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- * ------------------------------------------------------------------------------
- */
-package org.openmicroscopy.blitz
-
-import groovy.transform.CompileStatic
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.openmicroscopy.blitz.extensions.BlitzExtension
-
-@CompileStatic
-class BasePlugin implements Plugin<Project> {
-
-    public static final String EXTENSION_BLITZ = "blitz"
-
-    @Override
-    void apply(Project project) {
-        project.extensions.create(EXTENSION_BLITZ, BlitzExtension, project)
-    }
-
-}
