+++ src/main/groovy/org/openmicroscopy/blitz/BlitzIcePlugin.groovy
@@ -1,67 +1,78 @@
 package org.openmicroscopy.blitz
 
-
+import com.zeroc.gradle.icebuilder.slice.SlicePlugin
 import groovy.transform.CompileStatic
 import org.gradle.api.Action
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.Task
-import org.gradle.api.file.CopySpec
-import org.gradle.api.tasks.Copy
+import org.gradle.api.file.ProjectLayout
 import org.gradle.api.tasks.TaskProvider
+import org.gradle.api.tasks.bundling.Zip
+import org.gradle.api.tasks.util.PatternFilterable
 import org.openmicroscopy.blitz.extensions.IceExtension
+import org.openmicroscopy.tasks.IceDocsTask
+
+import javax.inject.Inject
 
 @CompileStatic
 class BlitzIcePlugin implements Plugin<Project> {
 
-    public static final String TASK_PROCESS_SLICE = "processSlice"
+    public static final String EXTENSION_ICE = "ice"
+
+    public static final String TASK_ZIP_ICEDOC = "zipIcedoc"
 
     public static final String TASK_COMPILE_ICEDOC = "compileIcedoc"
 
     private Project project
 
-    @Override
-    void apply(Project project) {
-        this.project = project
+    private IceExtension ice
 
-        // Apply the zeroc plugin
-        project.pluginManager.apply(IcePlugin)
+    private final ProjectLayout layout
 
-        project.pluginManager.apply(BlitzApiPlugin)
-
-        // Sort out ordering of tasks
-        registerProcessSliceTask()
-        configureTaskOrdering()
+    @Inject
+    BlitzIcePlugin(ProjectLayout layout) {
+        this.layout = layout
     }
 
-    TaskProvider<Copy> registerProcessSliceTask() {
-        project.tasks.register(TASK_PROCESS_SLICE, Copy, new Action<Copy>() {
+    @Override
+    void apply(Project project) {
+        // Apply the zeroc plugin
+        project.pluginManager.apply(SlicePlugin)
+
+        this.project = project
+
+        // Create ice extension
+        this.ice = project.extensions.create(EXTENSION_ICE, IceExtension, project)
+
+        // Register our tasks
+        registerIceDocsTask()
+        registerZipIceDocs()
+    }
+
+    TaskProvider<IceDocsTask> registerIceDocsTask() {
+        project.tasks.register(TASK_COMPILE_ICEDOC, IceDocsTask, new Action<IceDocsTask>() {
             @Override
-            void execute(Copy task) {
-                // Copy ice files generated in combinedToIce task to iceSrcDir
-                CopySpec combinedToIceSpec = project.copySpec()
-                combinedToIceSpec.from(project.tasks.named("combinedToIce"))
-                combinedToIceSpec.into("omero/model")
-
-                // Copy files located in project dir to iceSrcDir
-                CopySpec mainSpec = project.copySpec()
-                mainSpec.from(project.layout.projectDirectory.dir("src/main/slice"))
-
-                IceExtension ice = project.extensions.getByType(IceExtension,)
-                task.into(ice.iceSrcDir)
-                task.with(mainSpec)
-                task.with(combinedToIceSpec)
+            void execute(IceDocsTask task) {
+                task.source = project.fileTree(ice.iceSrcDir).matching(new Action<PatternFilterable>() {
+                    @Override
+                    void execute(PatternFilterable filter) {
+                        filter.include "**/*.ice"
+                    }
+                })
+                task.includeDirs.add(ice.iceSrcDir)
+                task.outputDir.set(ice.docsOutputDir)
             }
         })
     }
 
-    void configureTaskOrdering() {
-        TaskProvider<Task> processSlice = project.tasks.named(TASK_PROCESS_SLICE)
-
-        // Ice docs task depends on all ice files being present
-        project.tasks.named(TASK_COMPILE_ICEDOC).configure {
-            it.dependsOn(processSlice)
-        }
+    TaskProvider<Zip> registerZipIceDocs() {
+        project.tasks.register(TASK_ZIP_ICEDOC, Zip, new Action<Zip>() {
+            @Override
+            void execute(Zip zip) {
+                zip.archiveClassifier.set("icedoc")
+                zip.from(project.tasks.named(TASK_COMPILE_ICEDOC))
+            }
+        })
     }
 
 }
